Extract 2 classes from SearchService
Extract interface of SearchService.
And extract 2 classes from SearchService

 - Extract enum class of IndexType
 - Differentiate DocumentFactory from IndexType

As the package of IndexType changes,
fixed the class that referred to IndexType.
And correct the service reference setting of SonosHelper.
Because SonosHelper refers to the service entity, not auto inject.

DocumentFactory is heavily influenced by Lucene version.
These two classes have a major impact on index file management.
It is desirable to rebuild the index file
 at the time of design change due to requirements
 or implementation change of Lucene version upgrade.
