+++ airsonic-main/src/main/java/org/airsonic/player/service/SonosService.java
@@ -25,6 +25,7 @@ import org.airsonic.player.domain.AlbumListType;
 import org.airsonic.player.domain.MediaFile;
 import org.airsonic.player.domain.Playlist;
 import org.airsonic.player.domain.User;
+import org.airsonic.player.service.search.IndexType;
 import org.airsonic.player.service.sonos.SonosHelper;
 import org.airsonic.player.service.sonos.SonosServiceRegistration;
 import org.airsonic.player.service.sonos.SonosSoapFault;
@@ -264,13 +265,13 @@ public class SonosService implements SonosSoap {
     public SearchResponse search(Search parameters) {
         String id = parameters.getId();
 
-        SearchService.IndexType indexType;
+        IndexType indexType;
         if (ID_SEARCH_ARTISTS.equals(id)) {
-            indexType = SearchService.IndexType.ARTIST;
+            indexType = IndexType.ARTIST;
         } else if (ID_SEARCH_ALBUMS.equals(id)) {
-            indexType = SearchService.IndexType.ALBUM;
+            indexType = IndexType.ALBUM;
         } else if (ID_SEARCH_SONGS.equals(id)) {
-            indexType = SearchService.IndexType.SONG;
+            indexType = IndexType.SONG;
         } else {
             throw new IllegalArgumentException("Invalid search category: " + id);
         }
