+++ airsonic-main/src/main/java/org/airsonic/player/service/sonos/SonosHelper.java
@@ -27,6 +27,7 @@ import org.airsonic.player.controller.CoverArtController;
 import org.airsonic.player.dao.MediaFileDao;
 import org.airsonic.player.domain.*;
 import org.airsonic.player.service.*;
+import org.airsonic.player.service.search.IndexType;
 import org.airsonic.player.util.StringUtil;
 import org.airsonic.player.util.Util;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -523,7 +524,7 @@ public class SonosHelper {
         return Arrays.asList(artists, albums, songs);
     }
 
-    public MediaList forSearch(String query, int offset, int count, SearchService.IndexType indexType, String username, HttpServletRequest request) {
+    public MediaList forSearch(String query, int offset, int count, IndexType indexType, String username, HttpServletRequest request) {
 
         SearchCriteria searchCriteria = new SearchCriteria();
         searchCriteria.setCount(count);
