+++ airsonic-main/src/test/java/org/airsonic/player/service/search/SearchServiceTestCase.java
@@ -29,8 +29,8 @@ import org.airsonic.player.domain.SearchCriteria;
 import org.airsonic.player.domain.SearchResult;
 import org.airsonic.player.service.MediaScannerService;
 import org.airsonic.player.service.SearchService;
-import org.airsonic.player.service.SearchService.IndexType;
 import org.airsonic.player.service.SettingsService;
+import org.airsonic.player.service.search.IndexType;
 import org.airsonic.player.util.HomeRule;
 import org.junit.Assert;
 import org.junit.Before;
